refactor: refactor comment component
